# docs/swagger.yaml
basePath: /
definitions:
...
  models.AllHeirarchyRequest:
    properties:
      get_id:
        type: string
    type: object
  models.AllHeirarchyResponse:
    properties:
      branch_code:
        type: string
      branch_desc:
        type: string
      center_code:
        type: string
      center_desc:
        type: string
      hierarchy_id:
        type: string
      unit_code:
        type: string
      unit_desc:
        type: string
    type: object
...
  models.AllUserManagementRequest:
    properties:
      get_id:
        type: string
    type: object
  models.AllUserManagementResponse:
    properties:
      got_branch_desc:
        type: string
      got_check_status:
        type: string
      got_given_name:
        type: string
      got_inst_desc:
        type: string
      got_last_name:
        type: string
      got_middle_name:
        type: string
      got_role_name:
        type: string
      got_user_email:
        type: string
      got_user_name:
        type: string
      got_user_phone:
        type: string
      got_user_status:
        type: string
    type: object
...
  models.EditHeirarchyResponse:
    properties:
      update_hierarchy:
        type: string
    type: object
  models.EditHierarchyRequest:
    properties:
      get_branch_code:
        type: string
      get_branch_desc:
        type: string
      get_center_code:
        type: string
      get_center_desc:
        type: string
      get_hierarchy_id:
        type: string
      get_unit_code:
        type: string
      get_unit_desc:
        type: string
    type: object
...
  models.EditUserManagementRequest:
    properties:
      get_branch_desc:
        type: string
      get_check_status:
        type: string
      get_given_name:
        type: string
      get_inst_desc:
        type: string
      get_last_name:
        type: string
      get_middle_name:
        type: string
      get_role_name:
        type: string
      get_user_email:
        type: string
      get_user_id:
        type: string
      get_user_name:
        type: string
      get_user_phone:
        type: string
      get_user_status:
        type: string
    type: object
  models.EditUserManagementResponse:
    properties:
      update_usermanagement:
        type: string
    type: object
...
  models.HierarchyRequest:
    properties:
      b.branch_desc:
        type: string
      c.center_desc:
        type: string
      i.inst_desc:
        type: string
      u.unit_desc:
        type: string
    type: object
  models.HierarchyResponse:
    properties:
      branch_code:
        type: string
      branch_desc:
        type: string
      center_code:
        type: string
      center_desc:
        type: string
      hierarchy_id:
        type: string
      i.inst_desc:
        type: string
      unit_code:
        type: string
      unit_desc:
        type: string
    type: object
...
  models.ResponseModel:
    properties:
      data: {}
      message:
        type: string
      retcode:
        type: string
    type: object
...
  models.RolesManagementRequest:
    properties:
      role_name:
        type: string
    type: object
  models.RolesManagementResponse:
    properties:
      role_desc:
        type: string
      role_id:
        type: string
      role_name:
        type: string
    type: object
...
  models.UserManagementRequest:
    properties:
      branch_names:
        type: string
      check_status:
        type: string
      given_name:
        type: string
      last_name:
        type: string
      middle_name:
        type: string
      user_login:
        type: string
    type: object
  models.UserManagementResponse:
    properties:
      branch_names:
        type: string
      check_status:
        type: string
      given_name:
        type: string
      last_name:
        type: string
      middle_name:
        type: string
      roles:
        type: string
      user_id:
        type: string
      user_login:
        type: string
    type: object
...
host: 10.21.0.66:1111
info:
  contact:
    email: alex.dequilla@fortress-asya.com
    name: FDSAP-A.Dequilla
  description: KPLUS Webtool Swagger
  license: {}
  termsOfService: http://swagger.io/terms/
  title: Webtool
  version: "2.0"
paths:
...
  /edit_hierarchy/:
    post:
      consumes:
      - application/json
      description: Fetch User Data
      parameters:
      - description: EditHierarchy Input
        in: body
        name: edithierarchyInput
        required: true
        schema:
          $ref: '#/definitions/models.EditHierarchyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.EditHeirarchyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseModel'
      summary: Fetch User Data
      tags:
      - Webtool
...
  /edit_usermanagement/:
    post:
      consumes:
      - application/json
      description: Fetch User Data
      parameters:
      - description: EditUser Input
        in: body
        name: editUserInput
        required: true
        schema:
          $ref: '#/definitions/models.EditUserManagementRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.EditUserManagementResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseModel'
      summary: Fetch User Data
      tags:
      - Webtool
...
  /get_hierarchy/:
    post:
      consumes:
      - application/json
      description: Fetch User Data
      parameters:
      - description: Hierarchy Input
        in: body
        name: hierarchyInput
        required: true
        schema:
          $ref: '#/definitions/models.HierarchyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.HierarchyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseModel'
      summary: Fetch User Data
      tags:
      - Webtool
...
  /get_rolesmanagement/:
    post:
      consumes:
      - application/json
      description: Fetch User Data
      parameters:
      - description: Roles Input
        in: body
        name: rolesInput
        required: true
        schema:
          $ref: '#/definitions/models.RolesManagementRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.RolesManagementResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseModel'
      summary: Fetch User Data
      tags:
      - Webtool
...
  /get_usermanagement/:
    post:
      consumes:
      - application/json
      description: Fetch User Data
      parameters:
      - description: User Input
        in: body
        name: userInput
        required: true
        schema:
          $ref: '#/definitions/models.UserManagementRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserManagementResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseModel'
      summary: Fetch User Data
      tags:
      - Webtool
...
  /select_heirarchy/:
    post:
      consumes:
      - application/json
      description: Fetch User Data
      parameters:
      - description: AllHeirarchy Input
        in: body
        name: allHeirarchyInput
        required: true
        schema:
          $ref: '#/definitions/models.AllHeirarchyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AllHeirarchyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseModel'
      summary: Fetch User Data
      tags:
      - Webtool
...
  /select_usermanagement/:
    post:
      consumes:
      - application/json
      description: Fetch User Data
      parameters:
      - description: AllUser Input
        in: body
        name: allUserInput
        required: true
        schema:
          $ref: '#/definitions/models.AllUserManagementRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AllUserManagementResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseModel'
      summary: Fetch User Data
      tags:
      - Webtool
swagger: "2.0"